etl clinical: move `find_sample` to shared module
The `etl clinical` command uses the `find_sample` function which is not
specific to clinical records. Move this function to a shared module for
reuse by future commands.
